Add Windows ProcessListener
Check the version of operating system, and implement Windows version of ProcessListener.

+++ src/screenLocker/autoOpen/ProcessListener.java
@@ -3,14 +3,17 @@ package screenLocker.autoOpen;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.util.ArrayList;
 import java.util.List;
 
 import screenLocker.MyTimer;
+import screenLocker.loader.Loader;
 
 public class ProcessListener extends Thread {
 	private List<String> _blacklist;
 	
 	public ProcessListener() {
+		_blacklist = new ArrayList<>();
 	}
 
 	public void run() {
@@ -22,22 +25,43 @@ public class ProcessListener extends Thread {
 			} catch (InterruptedException e1) {
 				e1.printStackTrace();
 			}
-			_blacklist = MyTimer.BlackList();
+			_blacklist.add("pietty0400b14.exe");
+			_blacklist.add("GitHubDesktop.exe");
+			
+			//_blacklist = MyTimer.BlackList();
 
-			for (String item: _blacklist) {
-				ProcessBuilder pb = new ProcessBuilder("ps", "-C", item);
-				Process p;
-
-				try {
-					p = pb.start();
-					BufferedReader buf = new BufferedReader(new InputStreamReader(p.getInputStream()));
-					buf.readLine();
-					String res = buf.readLine();
-					if (res != null) {
-						Runtime.getRuntime().exec("kill `ps -C " + item + " -o pid=`");
+			if(Loader.IsLinux()) {
+				for (String item: _blacklist) {
+					ProcessBuilder pb = new ProcessBuilder("ps", "-C", item);
+					Process p;
+	
+					try {
+						p = pb.start();
+						BufferedReader buf = new BufferedReader(new InputStreamReader(p.getInputStream()));
+						buf.readLine();
+						String res = buf.readLine();
+						if (res != null) {
+							Runtime.getRuntime().exec("kill `ps -C " + item + " -o pid=`");
+						}
+					} catch (IOException e) {
+						e.printStackTrace();
 					}
-				} catch (IOException e) {
-					e.printStackTrace();
+				}
+			}
+			if(Loader.IsWindows()) {
+				for (String item: _blacklist) {
+					try {
+						Process p = Runtime.getRuntime().exec("tasklist"+ " /FI \"imagename eq "+ item+ "\"");
+						BufferedReader buf = new BufferedReader(new InputStreamReader(p.getInputStream()));
+						buf.readLine();
+						String res = buf.readLine();
+						
+						if (res != null) {
+							Runtime.getRuntime().exec("taskkill /f /im " + item);
+						}
+					} catch (IOException e) {
+						e.printStackTrace();
+					}
 				}
 			}
 		}
